sisu-test-release/requirements.txt
dj-database-url==0.5.0
Django==2.0.5
django-allauth==0.39.1
django-enumfields==0.10.1
django-heroku==0.3.1
django-ipware==2.1.0
django-social-widgets==0.4.0
gunicorn==19.9.0
numpy==1.15.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pygal==2.4.0
python-dateutil==2.7.3
python-decouple==3.1
python-http-client==3.1.0
pytz==2018.4
scikit-learn==0.20.0
scipy==1.1.0
sendgrid==5.6.0
six==1.11.0
whitenoise==4.1.2
